ensure getStorageCostEstimateV2 usageInBytes is an integer

+++ src/main/scala/org/broadinstitute/dsde/firecloud/model/Workspace.scala
@@ -98,7 +98,7 @@ case class RawlsGroupMemberList(userEmails: Option[Seq[String]] = None,
                                 subGroupNames: Option[Seq[String]] = None
 )
 
-case class WorkspaceStorageUsageAndCostEstimate(estimate: BigDecimal, usageInBytes: BigDecimal, lastUpdated: Instant)
+case class WorkspaceStorageUsageAndCostEstimate(estimate: BigDecimal, usageInBytes: BigInt, lastUpdated: Instant)
 
 case class WorkspaceId(workspaceId: UUID)
 case class WorkspaceIdResponse(workspace: WorkspaceId)
